+++ demo/cancer_multitask.R
@@ -1,9 +1,8 @@
 rm(list=ls())
 library(simone)
 data(cancer)
-attach(cancer)
 
-res.coop <- simone(expr, tasks=status)
+res.coop <- simone(cancer$expr, tasks=cancer$status)
 plot(res.coop, ask=FALSE)
 
 glist <- getNetwork(res.coop, "BIC")
@@ -11,4 +10,3 @@ plot(glist[[1]],glist[[2]])
 glist <- getNetwork(res.coop, 65)
 plot(glist[[1]],glist[[2]])
 
-detach(cancer)